Related term ES index speed, refs #13175
This commit contains an enhancement to speed up ES indexing of
information objects when there is a large number of terms in the terms
table.

- Subject, place and genre term ids and parent_ids are preloaded into an
array so a term's parent can be found directly using the term id as the
array index.

For each directly related term, parents are found by searching the term
array recursively instead of using the nested set ancestors query using
the LFT and RGT columns.

This change results in a large performance increase when indexing
information objects when testing with ~350K place and subject terms.

- Changed related terms query to use a subquery for performance. The old
query would take 12 seconds in some cases. The subquery technique is
faster in all cases.

- Load the terms lookup table straight from PDO using PDO::FETCH_KEY_PAIR
option. This removes the need to build the lookup table by hand.

Testing shows a performance increase when using PDO::FETCH_KEY_PAIR over
building the lookup table by hand.