Update multiFileUpload.js
Added an additional regular expression to the file renaming process to enable a more complex auto incrementing scheme. The original functionality is intact: 
With no user specified numbering scheme, sequential numbers are appended to the end of the filename string.
If the user specifies the %ddd% notation - then the numbers are padded out with leading zeros in the resultant string.

New functionality:
The user can specify the starting number for the incrementing scheme by putting it into the regex, so to start counting from 101 (instead of 1) the user would use the form: %ddd 101% - add a space after the placeholder markers and then instruct the method with the starting offset.

This pull request is the result of a user who needed to upload batches of files from different folders on their workstation, or at different times. The existing upload process would duplicate names. They were either manually editing names (in the UI, or with a CSV extract and import process) or pre-appending a digit before the %ddd% notation, but both solutions were not ideal.

This is a small change, and it may not be the best implementation. I do however think it is a useful and could be expanded for further niche naming systems in future versions. Like alphabetic, hexadecimal or perhaps a barcode or hashed serial number.